Refactor/server middlewares (#42)
* chore(tsconfig): rename @middlewares/* paths to @middlewares

* feat: add index.ts to middlewares folder

* refactor: replace export default with named export

* refactor: rename import @middlewares/[...] to @middlewares
