+++ server/src/middlewares/authorizationTwitchMiddleware.ts
@@ -5,7 +5,7 @@ import { AuthorizationTwitch } from "@types";
 import { createNewAuth } from "@services";
 import { clientId, clientSecret, redirectUrl } from "@configs";
 
-const authorizationTwitch = async (req: Request, res: Response, next: NextFunction) => {
+export const authorizationTwitch = async (req: Request, res: Response, next: NextFunction) => {
   const { code } = req.query as unknown as RequestQueryAuthorizationTwitch;
   const authRes = await retryWithCatch(() =>
     fetch("https://id.twitch.tv/oauth2/token", {
@@ -36,5 +36,3 @@ const authorizationTwitch = async (req: Request, res: Response, next: NextFuncti
     return next();
   }
 };
-
-export default authorizationTwitch;
