+++ server/src/routes/messageCategoriesRoute.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import checkSearchParams from "@middlewares/checkSearchParamsMiddleware";
+import { checkSearchParams, isParamObjectId } from "@middlewares";
 import {
   getMessageCategoriesList,
   editMessageCategoryById,
@@ -7,7 +7,6 @@ import {
   deleteMessageCategoryById,
   updateUsesCategoryById
 } from "@controllers/messageCategoriesController";
-import isParamObjectId from "@middlewares/isParamObjectIdMiddleware";
 
 const messageCategoriesRouter = Router();
