+++ server/src/routes/redemptionsRoute.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { getRedemptionsList } from "@controllers/redemptionsController";
-import checkSearchParams from "@middlewares/checkSearchParamsMiddleware";
+import { checkSearchParams } from "@middlewares";
 
 const redemptionsRouter = Router();
